requirements.txt
Flask==0.10.1
cassandra-driver==1.1.2
